Share fixture classes used by distance and closeness unit tests.
