+++ searchlib/src/vespa/searchlib/fef/test/labels.h
@@ -1,5 +1,7 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
+#pragma once
+
 #include <vespa/searchlib/fef/properties.h>
 #include <vespa/vespalib/stllike/asciistream.h>
