+++ searchlib/src/vespa/searchlib/test/features/CMakeLists.txt
@@ -1,0 +1,6 @@
+# Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+vespa_add_library(searchlib_searchlib_test_features
+    SOURCES
+    distance_closeness_fixture.cpp
+    DEPENDS
+)
