+++ searchlib/src/vespa/searchlib/test/features/distance_closeness_fixture.cpp
@@ -1,0 +1,10 @@
+// Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+
+#include "distance_closeness_fixture.h"
+
+namespace search::features::test {
+
+FeatureDumpFixture::~FeatureDumpFixture() = default;
+
+}
+
